\name{print.csFSSgr}
\alias{print.csFSSgr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print \code{csFSSgr} object.
}
\description{Prints out information on a \code{csFSSgr} object.
}
\usage{
\method{print}{csFSSgr}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{The \code{csFSSgr} object you wish to print.
}
  \item{\dots}{Other arguments.
}
}
\value{None
}
\references{
Cardinali, A. and Nason, G.P. (2012) Costationarity of Locally 
Stationary Time Series using \code{costat}.

Cardinali, A. and Nason, G.P. (2010) Costationarity of locally stationary
time series. \emph{J. Time Series Econometrics}, \bold{2}, Issue 2, Article 1.
}
\author{
Guy Nason
}

\seealso{\code{\link{plot.csFSSgr}}, \code{\link{summary.csFSSgr}}
}
\examples{
#
# The user should normally never need to use this function as the
# csFSSgr object is only ever internally produced and used.
#
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ts}
