\name{summary.csBiFunction}
\alias{summary.csBiFunction}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Summarize a \code{csBiFunction} object.
}
\description{Summarize a \code{csBiFunction} object.
}
\usage{
\method{summary}{csBiFunction}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{The object to summarize
}
  \item{\dots}{Other arguments
}
}
\value{None
}

\references{Cardinali, A. and Nason, Guy P. (2013) Costationarity of
	Locally Stationary Time Series Using costat.
	\emph{Journal of Statistical Software}, \bold{55}, Issue 1.

Cardinali, A. and Nason, G.P. (2010) Costationarity of locally stationary
time series. \emph{J. Time Series Econometrics}, \bold{2}, Issue 2, Article 1.
}
\author{
Guy Nason
}

\seealso{
\code{\link{plot.csBiFunction}}, \code{\link{print.csBiFunction}}
}
\examples{
#
# See example to print.csBiFunction, as this calls summary(.)
#
}
\keyword{ts}
