\name{summary.csFSSgr}
\alias{summary.csFSSgr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Summarize a \code{csFSSgr} object.
}
\description{Summarizes a \code{csFSSgr} object.
}
\usage{
\method{summary}{csFSSgr}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{The \code{csFSSgr} object you wish to summarize.
}
  \item{\dots}{Other arguments}
}
\value{None
}
\references{
Cardinali, A. and Nason, G.P. (2012) Costationarity of Locally 
Stationary Time Series using \code{costat}.

Cardinali, A. and Nason, G.P. (2010) Costationarity of locally stationary
time series. \emph{J. Time Series Econometrics}, \bold{2}, Issue 2, Article 1.
}
\author{
Guy Nason
}

\seealso{\code{\link{plot.csFSSgr}}, \code{\link{print.csFSSgr}}}
\examples{
#
# See example for print.csFSSgr which calls summary(.)
}
\keyword{ts}
